+++ network_automation/sdwan_ops/views.py
@@ -33,13 +33,30 @@ def vmanage_auth():
 def ops():
     return render_template(f"{TEMPLATE_DIR}/ops.html")
 
-@sdwan_ops.route("/snmp")
-def snmp():
+@sdwan_ops.route("/hostname")
+def hostname():
+    dev_results= []
     username = session.get("username")
     password = session.get("password")
     #username = "admin"
     #password = "C1sco12345"
-    auth_header = update_hostname.auth(VMANAGE_URL_VAR, username, password)
-    device_list = update_hostname.update_hostname(VMANAGE_URL_VAR, auth_header)
+    #auth_header = update_hostname.auth(VMANAGE_URL_VAR, username, password)
+    result = update_hostname.update_hostname(VMANAGE_URL_VAR, username, password)
 
-    return str(device_list)
+    summary = result["summary"]
+    for dev_result in result["data"]:
+        dev_data = (dev_result["host-name"], dev_result["status"])
+        dev_results.append(dev_data)
+
+    print(summary)
+    print("*" * 100)
+    print(dev_results)
+    return render_template(f"{TEMPLATE_DIR}/summary.html", result=result, summary=summary)
+
+@sdwan_ops.route("/success")
+def success():
+    return render_template(f"{TEMPLATE_DIR}/success.html")
+
+@sdwan_ops.route("/failure")
+def failure():
+    return render_template(f"{TEMPLATE_DIR}/failure.html")
