+++ network_automation/templates/sdwan_ops/summary.html
@@ -6,12 +6,10 @@
     </head>
     <body>
         <div class="alert alert-info" role="alert">
-            <h1>SD-WAN Operations</h1><br>
-            <h3>Data Input</h3><br>
-            <h5>Select Operation<br><br></h5>
-            <a class="btn btn-primary" href="{{url_for('sdwan_ops.hostname')}}" role="button">Update Hostname</a>
-            <a class="btn btn-primary" href="{{url_for('home')}}" role="button">Main Index</a>      
-        </div>
+        <h1>SDWAN OPS <br><br></h1>
+        <h4>Operation Summary: <br><br></h4> 
         <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
+        <a class="btn btn-primary" href="{{url_for('home')}}" role="button">Back To Home</a>
+        </div>
     </body>
 </html>
